serial with ale code
