+++ src/sensors/ManipulatorPickup.cpp
@@ -1,3 +1,5 @@
+//Alessandro -- please transfer any code for ManipulatorPickup.cpp into the sensors version of this file
+
 
 // Force Inclusions
 #include <Arduino.h>
@@ -10,13 +12,13 @@
 // Stepper Inclusions
 #include <UMS3.h>
 
+// serial
+#include "esp_sender2.h"
+#include <atomic>
+#include "util.h"
+
 // Force Definitions
 #define FORCE_SENSOR_PIN A3
-
-// Serial inclusion
-#include "esp_sender2.h"
-#include <util.h>
-#include <atomic>
 
 // Servo Definitions
 
@@ -39,22 +41,26 @@ int Step_Time = 0;
 long dt_Step = 2;
 bool wasHigh = true;
 int Step_Up = 0;
-UMS3 ums3;
+// UMS3 ums3;
 
 bool navigating = false;
-bool pickup = true;
-bool dropoff = false;
+bool pickupManip = true;
+bool dropoffManip = false;
+bool activate = false;
 
-std::atomic<int> task{-1}; //1 = pickup, 2 = dropoff
+// PySerial Definitions
+int incomingByte = 0;
+
+std::atomic<int> task{-1};
 
 void open_gripper(){
-    Serial.println("Opening...");
+    //Serial.println("Opening...");
     myservo.write(150); 
     delay(2000);
 }
 
 void close_gripper(){
-    Serial.println("Closing...");
+    //Serial.println("Closing...");
     myservo.write(50);
     delay(2000);
 }
@@ -83,32 +89,46 @@ void raise_stepper(){
     }
 }
 
-void setup() {
-  Serial.begin();
+void manipulatorSetup() {
+  //wireless
+  // Serial.begin();
+
+  // ums3.begin();
+
+  // //initPeripherals();
+  // //initRotary();  
+  // Serial.println("Starting!");
+  // delay(1000);
+  // initReceiver();
 
   // Force Sensor
   // set the ADC attenuation to 11 dB (up to ~3.3V input)
   analogSetAttenuation(ADC_11db);
-  Serial.println("Force Sensor Activated");
+  //Serial.println("Force Sensor Activated");
 
   // Servo
-  Serial.println("Starting Servo Test");
+  //Serial.println("Starting Servo Test");
 
   myservo.setPeriodHertz(60);    // standard 50 hz servo
   myservo.attach(servoPin, 500, 2500); // attaches the servo on pin 18 to the servo object
 
-  Serial.println("Servo Activated");
+  //Serial.println("Servo Activated");
 
   // Stepper
-  ums3.begin();
+  // ums3.begin();
   pinMode(15, OUTPUT); //step pin
   pinMode(16, OUTPUT); // directinon pinMatrixInAttach
   
-  Serial.println("Stepper Activated");
+  //Serial.println("Stepper Activated");
   delay(10);
 }
 
-void loop() {
+void manipulatorLoop() {
+  // printData();
+  // if(data.swch1){
+  //   activate = true;
+  //   delay(1000);
+  // }
   // testing 
    /*
   {  
@@ -184,11 +204,20 @@ void loop() {
 }
   */
     // use pyserial here in the loop to see when it's time to go into man mode (the loop automatically takes us out of man mode)
+    // if (Serial.available() > 0){
+    //   incomingByte = Serial.read();
+
+    //   Serial.println("I received his good word: ");
+    //   Serial.println(incomingByte, DEC);
+    // }
+
+    /*Serial.println("Manipulato rPrint Test");
+
 
     // glory to his kingdom
-    /*while(!navigating){
+    if(activate){
 
-        if(pickup) {
+        if(pickupManip) {
         // make sure the stepper is as low as possible and the servo is open as much as possible 
         // hope();
         // start the servo closing until the force sensor reads a nonzero value
@@ -197,27 +226,51 @@ void loop() {
         // force_feedback();
         // once this is done, lift 
         raise_stepper();
-        pickup = false;
-        dropoff = true; 
-        } else if(dropoff){
+        pickupManip = false;
+        dropoffManip = true; 
+        } else if(dropoffManip){
         // move the stepper motor down as low as possible
         lower_stepper();
         // open the servo as far as possible
         open_gripper();
-        dropoff = false;
-        pickup = true;
-        } else {
-            navigating = true;
-        }
-    }*/
+        dropoffManip = false;
+        pickupManip = true;
 
-    EVERY_N_MILLIS(500) { //TODO: finetune this delay
+        delay(2000);
+
+    }
+    activate = false;
+    delay(1000);
+  }*/
+
+
+    EVERY_N_MILLIS(200) { //TODO: finetune this delay
       SensorData d = loopComm();
       if (d.task != -1) { 
           task.store(d.task); //1 = pickup, 2 = dropoff
-          //TODO: for Alessandro -- execute the task here, probably.
-            //call sendToJetson() when you're done (so the state machine knows you're done)
-      }
+        
+
+          if(d.task == 1) {
+            // make sure the stepper is as low as possible and the servo is open as much as possible 
+            // hope();
+            // start the servo closing until the force sensor reads a nonzero value
+            close_gripper();
+            // when the force sensor begins to read, set a setpoint (prob the max??) and control it open loop? 
+            // force_feedback();
+            // once this is done, lift 
+            raise_stepper();
+
+            sendToJetson(); //let jetson know we're ending task
+          } else if(d.task == 2){
+            // move the stepper motor down as low as possible
+            lower_stepper();
+            // open the servo as far as possible
+            open_gripper();
+    
+            delay(2000); //TODO: Alessandro -- do we need this?
+            sendToJetson();  //let jetson know we're ending task
+        
+          }
     }
 
 }
